Updated some controllers --Added read me file

+++ WebApplication3/Models/UserDetails.cs
@@ -5,12 +5,12 @@ namespace WebApplication3.Models
 {
     public partial class UserDetails
     {
-        public long Id { get; set; }
+        public decimal Id { get; set; }
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Gender { get; set; }
-        public long? LoginId { get; set; }
-        public long? EducationId { get; set; }
+        public decimal? LoginId { get; set; }
+        public decimal? EducationId { get; set; }
 
         public virtual Education Education { get; set; }
         public virtual UserLogin Login { get; set; }
